Don't rebuild links for every source change
Theoretically it is the right thing to do, since we might add a new links
target that has been referenced somewhere else. In practice it means every html
file is rebuild for every source file change.

With this change, the individual html files no longer depend on the
links file. This means that for a source change, only the links file and
the corresponding html file is rebuilt. This means new links are picked
up in the changed file only.

To make sure all links resolve, run `make clean; make`.